Add files via upload
I found this page on Instagram and saw it was lacking responsive styling, so i decided to add it for the home page, from what I've seen of the website this can be applied around the site fairly simply, mainly with changes to the images and CSS. I remade the icons and text logo image as svgs so they are cleaner when scaled too. I also refactored the HTML to remove outdated syntax, moved sections into nav and footer blocks and made the background of the light mode slightly darker.